+++ watch/tasks/fusion/datamodules/spacetime_grid_builder.py
@@ -246,7 +246,8 @@ def sample_video_spacetime_targets(dset, window_dims, window_overlap=0.0,
     parts = set(time_sampling.split('+'))
     affinity_type_parts = parts & {
         'hard', 'hardish', 'contiguous', 'soft2', 'soft', 'hardish2',
-        'hardish3', 'soft2-contiguous-hardish3',
+        'hardish3', 'soft2-contiguous-hardish3', 'uniform',
+        'uniform-soft2-contiguous-hardish3',
     }
     update_rule_parts = parts & {'distribute', 'pairwise'}
     unknown = (parts - affinity_type_parts) - update_rule_parts